- fixed: Assignment of GetLightLevel to GLWall::lightlevel must be clamped. This was done incorrectly in the last SVN version which clamped the result of GetLightLevel, causing problems elsewhere.

+++ src/gl/scene/gl_wall.h
@@ -118,7 +118,7 @@ public:
 	
 	fixed_t viewdistance;
 
-	BYTE lightlevel;
+	int lightlevel;
 	BYTE type;
 	BYTE flags;
 	short rellight;
